+++ spikeinterface_gui/isiview.py
@@ -1,48 +1,60 @@
-from .myqt import QT
-import pyqtgraph as pg
+# from .myqt import QT
+# import pyqtgraph as pg
 
-import numpy as np
-import pandas as pd
+# import numpy as np
+# import pandas as pd
 
-from .base import WidgetBase
+# from .base import WidgetBase
 
-
-
-class MyViewBox(pg.ViewBox):
-    doubleclicked = QT.pyqtSignal()
-    def mouseDoubleClickEvent(self, ev):
-        self.doubleclicked.emit()
-        ev.accept()
-    def raiseContextMenu(self, ev):
-        #for some reasons enableMenu=False is not taken (bug ????)
-        pass
+from .view_base import ViewBase
 
 
 
-class ISIView(WidgetBase):
+
+# class MyViewBox(pg.ViewBox):
+#     doubleclicked = QT.pyqtSignal()
+#     def mouseDoubleClickEvent(self, ev):
+#         self.doubleclicked.emit()
+#         ev.accept()
+#     def raiseContextMenu(self, ev):
+#         #for some reasons enableMenu=False is not taken (bug ????)
+#         pass
+
+
+
+class ISIView(ViewBase):
+    _supported_backend = ['qt']
     _settings = [
                 {'name': 'window_ms', 'type': 'float', 'value' : 50. },
                 {'name': 'bin_ms', 'type': 'float', 'value' : 1.0 },
         ]
     _need_compute = True
-    
-    def __init__(self, controller=None, parent=None):
-        WidgetBase.__init__(self, parent=parent, controller=controller)
+
+
+    def __init__(self, controller=None, parent=None, backend="qt"):
+        ViewBase.__init__(self, controller=controller, parent=parent,  backend=backend)
+
+    # def __init__(self, controller=None, parent=None):
+    #     WidgetBase.__init__(self, parent=parent, controller=controller)
+
+
+    def _make_layout_qt(self):
+        import pyqtgraph as pg
+        from .myqt import QT
+        from .utils_qt import ViewBoxHandlingDoubleClick
         
+
+
         self.layout = QT.QVBoxLayout()
-        self.setLayout(self.layout)
+        # self.setLayout(self.layout)
         
         self.graphicsview = pg.GraphicsView()
         self.layout.addWidget(self.graphicsview)
         
         self.isi_histograms, self.isi_bins = self.controller.get_isi_histograms()
 
-        self.initialize_plot()
-        
-
-    def initialize_plot(self):
-        self.viewBox = MyViewBox()
-        self.viewBox.doubleclicked.connect(self.open_settings)
+        self.viewBox = ViewBoxHandlingDoubleClick()
+        # self.viewBox.doubleclicked.connect(self.open_settings)
         
         self.plot = pg.PlotItem(viewBox=self.viewBox)
         self.graphicsview.setCentralItem(self.plot)
@@ -51,14 +63,16 @@ class ISIView(WidgetBase):
 
     def compute(self):
         self.isi_histograms, self.isi_bins = self.controller.compute_isi_histograms(
-                self.params['window_ms'],  self.params['bin_ms'])
+                self.settings['window_ms'],  self.settings['bin_ms'])
         self.refresh()
 
     def on_params_changed(self):
         self.isi_histograms, self.isi_bins = None, None
         self.refresh()
 
-    def _refresh(self):
+    def _refresh_qt(self):
+        import pyqtgraph as pg
+        
         self.plot.clear()
         if self.isi_histograms is None:
             return
@@ -70,7 +84,7 @@ class ISIView(WidgetBase):
 
             isi = self.isi_histograms[unit_index, :]
             
-            qcolor = self.controller.qcolors[unit_id]
+            qcolor = self.get_unit_color(unit_id)
             curve = pg.PlotCurveItem(self.isi_bins[:-1], isi, pen=pg.mkPen(qcolor, width=3))
             self.plot.addItem(curve)
