+++ spikeinterface_gui/tests/test_mainwindow.py
@@ -82,8 +82,8 @@ def test_mainwindow(start_qt_app=False, verbose=True, curation=False, only_some_
     else:
         run_mainwindow(
             analyzer,
-            # backend="qt",
-            backend="panel",
+            backend="qt",
+            # backend="panel",
             start_qt_app=start_qt_app, verbose=verbose,
             curation=curation, curation_dict=curation_dict, 
             displayed_unit_properties=None,
